Filter out hits from clients that had very short reads.
In particular it seems there was one specific client that was repeatedly
downloading the first 76032 bytes of the raspi images, which was not a real
download and was completely swamping the statistics for the downloads.
Eventually this client went away, but in general we should avoid counting
incomplete downloads in our stats - and since we can't correlate downloads
by IP address in the logs we have, this is a good approximation.